+++ CHANGELOG.md
@@ -70,6 +70,10 @@ git tag -a "版本" -m "release pmj: 版本"
 - 修復：網頁 license
 - 新增：［pmj_server::gui］發送房間Root訊息
 
+**2026-9-10**
+
+- 更改：合并 `copy_website_dep.py` 到 `build_website.py`
+
 ## [v0.1.1] - 2026-8-17
 
 修復客戶端錯誤，並改進訊息顯示。